- Added Login CommandClass - Removed Strategy - Splitted strategy_spec to input_spec & authenticate_spec
