--- app/domain/authentication/strategy.rb
@@ -1,131 +1,0 @@
-# frozen_string_literal: true
-
-module Authentication
-
-  # - Runs security checks
-  # - Finds the appropriate authenticator
-  # - Validates credentials against
-  # - Returns a new token
-  class Strategy < ::Dry::Struct
-
-    AuthenticatorNotFound = ::Util::ErrorClass.new(
-      "'{0}' wasn't in the available authenticators")
-    InvalidCredentials = ::Util::ErrorClass.new(
-      "Invalid credentials")
-    InvalidOrigin = ::Util::ErrorClass.new(
-      "Invalid origin")
-
-
-    class Input < ::Dry::Struct
-      attribute :authenticator_name, ::Types::NonEmptyString
-      attribute :service_id, ::Types::NonEmptyString.optional
-      attribute :account, ::Types::NonEmptyString
-      attribute :username, ::Types::NonEmptyString.optional
-      attribute :password, ::Types::String
-      attribute :origin, ::Types::NonEmptyString
-      attribute :request, ::Types::Any.optional # for k8s authenticator
-
-      # Convert this Input to an Security::AccessRequest
-      #
-      def to_access_request(env)
-        ::Authentication::Security::AccessRequest.new(
-          webservice: webservice,
-          whitelisted_webservices: ::Authentication::Webservices.from_string(
-            account, env['CONJUR_AUTHENTICATORS'] ||
-            Authentication::Common.default_authenticator_name
-          ),
-          user_id: username
-        )
-      end
-
-      # Creates a copy of this object with the attributes updated by those
-      # specified in hash
-      #
-      def update(hash)
-        self.class.new(to_hash.merge(hash))
-      end
-
-      def webservice
-        @webservice ||= ::Authentication::Webservice.new(
-          account: account,
-          authenticator_name: authenticator_name,
-          service_id: service_id
-        )
-      end
-    end
-
-    # required constructor parameters
-    #
-    attribute :authenticators, ::Types::Hash
-
-    # optional constructor parameters
-    #
-    attribute :security, ::Types::Any.default {::Authentication::Security.new}
-    attribute :env, ::Types::Any.default(ENV)
-    attribute :token_factory, ::Types::Any.default {TokenFactory.new}
-    attribute :role_cls, ::Types::Any.default {::Role}
-    attribute :audit_log, ::Types::Any.default {AuditLog}
-    attribute :oidc_client_class, ::Types::Any.default {AuthnOidc::OidcClient}
-
-    def login(input)
-      authenticator = authenticators[input.authenticator_name]
-
-      validate_authenticator_exists(input, authenticator)
-      validate_security(input)
-
-      key = authenticator.login(input)
-      raise InvalidCredentials unless key
-
-      audit_success(input)
-      new_login(input, key)
-    rescue => err
-      audit_failure(input, err)
-      raise err
-    end
-
-    private
-
-    def audit_success(input)
-      audit_log.record_authn_event(
-        role: role(input.username, input.account),
-        webservice_id: input.webservice.resource_id,
-        authenticator_name: input.authenticator_name,
-        success: true
-      )
-    end
-
-    def audit_failure(input, err)
-      audit_log.record_authn_event(
-        role: role(input.username, input.account),
-        webservice_id: input.webservice.resource_id,
-        authenticator_name: input.authenticator_name,
-        success: false,
-        message: err.message
-      )
-    end
-
-    def role(username, account)
-      role_cls.by_login(username, account: account)
-    end
-
-    def validate_authenticator_exists(input, authenticator)
-      raise AuthenticatorNotFound, input.authenticator_name unless authenticator
-    end
-
-    def validate_security(input)
-      security.validate(input.to_access_request(env))
-    end
-
-    def validate_origin(input)
-      authn_role = role(input.username, input.account)
-      raise InvalidOrigin unless authn_role.valid_origin?(input.origin)
-    end
-
-    def new_login(input, key)
-      LoginResponse.new(
-        role_id: role(input.username, input.account).id,
-        authentication_key: key
-      )
-    end
-  end
-end
